api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.41
alembic==1.16.4
PyJWT==2.10.1
httpx==0.28.1
